Moved testcases to integration tests
