+++ python-sdk/tests/databases/test_base_database.py
@@ -1,5 +1,4 @@
 import pathlib
-from unittest import mock
 
 import pytest
 from pandas import DataFrame
@@ -8,8 +7,7 @@ from astro.constants import FileType
 from astro.databases import create_database
 from astro.databases.base import BaseDatabase
 from astro.files import File
-from astro.settings import SCHEMA
-from astro.table import BaseTable, Metadata, Table
+from astro.table import BaseTable, Table
 
 CWD = pathlib.Path(__file__).parent
 
@@ -98,25 +96,3 @@ def test_subclass_missing_append_table_raises_exception():
     target_table = Table()
     with pytest.raises(NotImplementedError):
         db.append_table(source_table, target_table, source_to_target_columns_map={})
-
-
-@mock.patch("astro.databases.base.BaseDatabase.drop_table")
-@mock.patch("astro.databases.base.BaseDatabase.create_schema_if_needed")
-@mock.patch("astro.databases.base.BaseDatabase.create_table")
-@mock.patch("astro.databases.base.BaseDatabase.load_file_to_table_natively_with_fallback")
-@mock.patch("astro.databases.base.resolve_file_path_pattern")
-def test_load_file_calls_resolve_file_path_pattern_with_filetype(
-    resolve_file_path_pattern,
-    load_file_to_table_natively_with_fallback,
-    create_table,
-    create_schema_if_needed,
-    drop_table,
-):
-    resolve_file_path_pattern.return_value = [File(path="S3://somebucket/test.csv")]
-    database = create_database("google_cloud_default")
-    database.load_file_to_table(
-        input_file=File(path="S3://somebucket/test.csv"),
-        output_table=Table(conn_id="google_cloud_default", metadata=Metadata(schema=SCHEMA)),
-        use_native_support=True,
-    )
-    assert resolve_file_path_pattern.call_args.kwargs["filetype"] == FileType.CSV
